fix: keep @kolu/surface-nix-host off the agent's runtime path (+ boot guard)
The agent runs raw .ts via bun, loading drishti-common -> common/surface.ts.
That AGENT-SHARED module imported @kolu/surface-nix-host (mirroredSurface +
browserSurface + the DEFAULT_CONNECTION value re-export), but the agent's scoped
build (nix/packages/drishti-agent, issue #38) deliberately hydrates only
@kolu/surface — so every remote agent died at import with
'Cannot find module @kolu/surface-nix-host/connection', flapping the connection
copying -> disconnected on every reconnect.

Mirrors kolu's split (terminalWorkspaceSurface carries no surface-nix-host dep):
- common/surface.ts is now pure base — NO @kolu/surface-nix-host import.
- NEW common/browser.ts (the drishti-common/browser subpath) holds browserSurface
  + the connection type/DEFAULT_CONNECTION re-exports; only the parent re-serve +
  the client import it. The agent imports drishti-common (surface.ts) and never
  reaches it.
- App.tsx / TabStrip.tsx / connectionColors.ts / wire.ts / router.ts: connection
  bits now import from drishti-common/browser.

Why CI missed it: typecheck resolves @kolu/* from the dev node_modules (has
everything) and 'nix build' only COPIES the agent tree — no lane ever EXECUTED
the agent. New 'agent-boots' CI lane builds + boots the agent on both platforms;
it fails with the exact 'Cannot find module' on a stray import (verified
adversarially) and passes now.

Re-pinned kolu -> 4e19d28fb.
