+++ packages/common/package.json
@@ -3,11 +3,12 @@
   "version": "0.1.0",
   "private": true,
   "type": "module",
-  "//": "Deps are what the wire contract needs at runtime: surface.ts uses zod and @kolu/surface/define (which pulls @orpc/contract). @kolu/surface is hydrated from the Nix store, not bun.lock, so its support deps must be declared by consumers — here for the contract, and in drishti-agent for the server/peer-server it serves.",
+  "//": "Deps are what the wire contract needs at runtime: surface.ts uses zod and @kolu/surface/define (which pulls @orpc/contract). @kolu/surface is hydrated from the Nix store, not bun.lock, so its support deps must be declared by consumers \u2014 here for the contract, and in drishti-agent for the server/peer-server it serves.",
   "main": "./src/surface.ts",
   "types": "./src/surface.ts",
   "exports": {
-    ".": "./src/surface.ts"
+    ".": "./src/surface.ts",
+    "./browser": "./src/browser.ts"
   },
   "scripts": {
     "typecheck": "tsc --noEmit"
